requirements.txt
numpy
scikit-learn
torch
torchvision
tqdm
ray[tune]
plotly
pandas
matplotlib
tensorboard
hydra-core

